Implementing the quick query aggregator.
